05-plataformas-de-datos/pipelines/claude-pipeline/pipeline/assets/ingestion/requirements.txt
pandas==2.2.0
requests==2.31.0
pyarrow==15.0.0
python-dateutil==2.8.2
